[ch452] starting on the tests
