+++ hermes-integration-test-app/grails-app/assets/stylesheets/mobile.css
@@ -28,7 +28,7 @@
         overflow: hidden;
     }
 
-    /* slightly smaller margin around content body */
+    /* slightly smaller margin around content responseBody */
     fieldset,
     .property-list {
         padding: 0.3em 1em 1em;
