+++ hermes-plugin/grails-app/services/com/atlasgenetics/hermes/message/MessageSenderService.groovy
@@ -1,6 +1,6 @@
 package com.atlasgenetics.hermes.message
 
-import com.atlasgenetics.hermes.response.HttpResponse
+import com.atlasgenetics.hermes.response.HermesResponseWrapper
 import grails.gorm.transactions.Transactional
 
 /**
@@ -14,25 +14,28 @@ class MessageSenderService {
     def failedMessageManagerService
     def httpService
 
-    HttpResponse sendNewMessage(MessageCommand message) {
-        HttpResponse response = httpService.makeRequest(message)
+    HermesResponseWrapper sendNewMessage(MessageCommand message) {
+        HermesResponseWrapper response = httpService.makeRequest(message)
         if (response.failed) {
             FailedMessage failedMessage = failedMessageManagerService.createFailedMessage(message, response.statusCode)
             if (!failedMessage.invalid) {
                 return retryFailedMessage(failedMessage, message)
+            } else {
+                response.failedMessageId = failedMessage.id
             }
         }
         return response
     }
 
-    HttpResponse retryFailedMessage(FailedMessage message, MessageCommand messageCommand = null) {
+    HermesResponseWrapper retryFailedMessage(FailedMessage message, MessageCommand messageCommand = null) {
         if (!message.invalid) {
             if (!messageCommand) messageCommand = new MessageCommand(message.messageData)
-            HttpResponse response = httpService.retryMessage(messageCommand, message.statusCode)
+            HermesResponseWrapper response = httpService.retryMessage(messageCommand, message.statusCode)
             if (response.succeeded) {
                 failedMessageManagerService.purgeMessage(message)
             } else {
                 failedMessageManagerService.completeFailedRetryProcess(message, response.statusCode)
+                response.failedMessageId = message.id
             }
             return response
         } else {
